Refactor: Move `jsx` related function to `jsx.js`
